+++ noodles/src/lib.rs
@@ -2,7 +2,7 @@
 
 //! **noodles** is a library for handling various bioinformatics file formats. It currently
 //! includes readers and writers for BAM 1.6, BCF 2.2, BED, BGZF, CRAM 3.0/3.1, CSI, FASTA, FASTQ,
-//! GFF3, GTF 2.2, SAM 1.6, tabix, and VCF 4.3.
+//! GFF3, GTF 2.2, SAM 1.6, tabix, and VCF 4.3/4.4.
 
 #[cfg(feature = "bam")]
 #[doc(inline)]